% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orthoIDs.R
\name{get.orthoIDs}
\alias{get.orthoIDs}
\title{Get ortholog gene IDs from Ensembl}
\usage{
get.orthoIDs(genes = NULL, query_species = "Homo sapiens",
  input = "external_gene_name", set = "all",
  host = "www.ensembl.org", check = FALSE, online = FALSE)
}
\arguments{
\item{genes}{A character vector of input genes. By default, Human gene names.}

\item{query_species}{A character indicating the query species name from which we will search orthologs. Defaults to Homo sapiens.}

\item{input}{A character indicating the type of gene identifiers. By default, gene names (external_gene_names). See \code{biomaRt::listFilters()} for more details.}

\item{set}{A character vector indicating names of species to search for orthologs. By default, all available species. However, I recommend to limit this to a number of species of interest.}

\item{host}{A character specifying the Ensembl mirror host. Defaults to the European mirror.}

\item{check}{A logical. If TRUE, it returns the Ensembl dictionary and stops. Useful for correcting mispelled species names}

\item{online}{A logical indicating if the dictionary should be created from online resources. If TRUE, it will run \code{get.dict}. Useful to get the most up-to-date dictionary, albeit slower. If FALSE, it will import dict dataset. Updated on April 2019. Defults to FALSE.}
}
\value{
A data frame containing Ensembl IDs and species names, among other information. If \code{check = TRUE}, it returns the Ensembl dictionary
}
\description{
Get Ensembl IDs of ortholog genes in desired species, that can subsequently used to retrieve their sequences using \code{get.seq} function.
}
\examples{

IDs <- get.orthoIDs(genes = c("APOE", "HTR3A", "PAX9"), set = c("Mus musculus", "Felis catus"))

}
