% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seq.R
\name{get.seq}
\alias{get.seq}
\title{Get sequences from Ensembl IDs}
\usage{
get.seq(IDs, seqtype = NULL, input_type = "ensembl_gene_id",
  longest = TRUE, host = "www.ensembl.org", online = FALSE)
}
\arguments{
\item{IDs}{A character vector containing the Ensembl identifiers, consistent with input_type}

\item{seqtype}{Type of output sequence, coding or protein ("coding", "peptide", respectively). See \code{biomaRt::listAttributes()} for more details.}

\item{input_type}{A character indicating the input type. By default Ensembl gene IDs (ensembl_gene_id). See \code{biomaRt::listFilters} for more details.}

\item{longest}{A logical indicating whether, in case a gene ID has several available transcripts, only the longest one should be kept. If FALSE, it will keep all available transcript sequences.}

\item{host}{A character indicating the Ensembl mirror host. Defaults to the European mirror.}

\item{online}{A logical indicating if the dictionary should be created from online resources. If TRUE, it will run \code{get.dict}. Useful to get the most up-to-date dictionary, albeit slower. If FALSE, it will import dict dataset. Updated on April 2019. Defults to FALSE.}
}
\value{
A data.frame with the desired sequences, corresponding IDs, and species names, among others
}
\description{
This function retrieves the coding or protein sequence from a list of identifiers.
}
\examples{
Seqfile <- get.seq(IDs = c("ENSG00000166736", "ENSMUSG00000032269"), seqtype = "peptide", host = "uswest.ensembl.org")
}
